+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.5.0
 
 - Support proto v0.51 changes and `.protolock` files.
+- Fixed the proto/moon versions not being available in the post-run cache key.
 - Updated dependencies.
 
 # 0.4.1